Reorganized Admin dashboard and components
